+++ pegasus/sites.v3/code.org/public/cms-demo.haml
@@ -71,9 +71,6 @@ video_player: true
     .quotes-wrapper
       = view :"ai_landing_page/ai_carousel_quotes"
 
-= view :swiper_page_cms_demo
-
-
