--- pegasus/sites.v3/code.org/views/swiper_page_cms_demo.haml
@@ -1,2 +1,0 @@
-// Adds Swiper to code.org/ai
-%script{src: webpack_asset_path('js/code.org/views/swiper_page_cms_demo.js')}
